store: Refactor clear_stale_call_cache
Refactor the code to separate the gymnastics caused by the two separate
storage schemes from the overall control loop

+++ graph/src/components/store/traits.rs
@@ -660,8 +660,8 @@ pub trait ChainStore: ChainHeadStore {
     /// Clears stale call cache entries for the given TTL in days.
     async fn clear_stale_call_cache(
         &self,
-        ttl_days: i32,
-        ttl_max_contracts: Option<i64>,
+        ttl_days: usize,
+        ttl_max_contracts: Option<usize>,
     ) -> Result<(), Error>;
 
     /// Return the chain identifier for this store.
